+++ packages/jimp/package.json
@@ -68,6 +68,7 @@
     "@jimp/plugin-blit": "0.3.5",
     "@jimp/plugin-rotate": "0.3.5",
     "@jimp/plugin-print": "0.3.5",
+    "@jimp/plugin-blur": "0.3.5",
     "@jimp/plugin-crop": "0.3.5",
     "@jimp/plugin-autocrop": "0.3.5"
   },
