+++ packages/plugin-blur/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-blur
+
+blur an image.
