+++ packages/plugin-blur/src/blur-tables.js
